New task execution framework (#107)
* misc: Upgrade navvy

* Refactor relation between task and cmd

* task: Add support for upload local file

* task/types: Add check before get value

* task/copy: Implement copy from stram/file to object

* task/common: Move mulptipart and object operation to common

* task: Split copy into copy_file and copy_stream

* task/utils: Remove not used funcs

* task: Rename FilePath to Path

* task: Generate task definitions

* task: Auto generate inherited value code

* task: Add copy root task

* task: Add ParseTowArgs and SetupStorage

* task: Add test for file task

* task/types: Add more clear info for get invalid value

* task/common: Add md5sum test

* task/common: Add test for multipart

* travis: Use go 1.13 instead

* task/copy_file: Add test for copy file tasks

* task/copy_stream: Add test for copy stream task

* utils: Add stream input support

* action: Remove not used tests

* task/utils: Move PraseInput into task

* task/tools: Format input json

* task: Add requirnment for dependent task

* task/tools: Refactor template

* task/tools: Do not add name in json file

* Fix conflict

+++ task/utils/random.go
@@ -28,7 +28,7 @@ func (r *randbo) Read(p []byte) (n int, err error) {
 	todo := len(p)
 	offset := 0
 	for {
-		val := int64(r.Int63())
+		val := r.Int63()
 		for i := 0; i < 7; i++ {
 			p[offset] = byte(val)
 			todo--
